Student/Professor

+++ People/Member.java
@@ -32,7 +32,7 @@ public class Member extends Person {
         this.numOfBooks=numOfBooks;
     }
 
-    public float getNumOfBooks(){
+    public int getNumOfBooks(){
         return numOfBooks;
     }
     public void printInfo(){
